Improve WASM SQL editor: case-insensitive columns, selection run, live status logs
Wrap Iceberg tables with lowercase column schemas so JOINs work with unquoted ids.
Run only selected SQL from the editor; stream idb_query progress in the status bar.

+++ iceberg-db-rs/crates/idb-wasm/src/lib.rs
@@ -201,7 +201,7 @@ fn truncate_log(s: &str, max: usize) -> String {
 #[derive(Serialize)]
 struct QueryResponse {
     row_count: usize,
-    elapsed_ms: u128,
+    elapsed_ms: u64,
     columns: Vec<ColumnDto>,
     rows: Vec<Vec<String>>,
     text: String,
